+++ app/Models/AcademicDiscipline.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class AcademicDiscipline extends Model
 {
@@ -12,5 +13,10 @@ class AcademicDiscipline extends Model
     public function parent(): BelongsTo
     {
         return $this->belongsTo(self::class, 'parent_id');
+    }
+
+    public function specializations(): HasMany
+    {
+        return $this->hasMany(self::class, 'parent_id');
     }
 }